fix test

+++ api/handler_events_test.go
@@ -123,7 +123,7 @@ func TestHandler_Events(t *testing.T) {
 								},
 								"url": "http://localhost/foo",
 							},
-							"response": map[string]interface{}{"body": "", "size": 0, "statusCode": 0}},
+							"response": map[string]interface{}{"body": "", "size": float64(0), "statusCode": float64(0)}},
 							m[0]["data"])
 					}))
 			},
